[move/#283] Notification 모델 패키지 위치 변경 (domain -> data)
