+++ app/src/main/java/com/smashing/app/data/repository/api/NotificationRepository.kt
@@ -2,7 +2,7 @@ package com.smashing.app.data.repository.api
 
 import com.smashing.app.data.model.cursor.CursorPage
 import com.smashing.app.data.model.notification.NotificationSportMatch
-import com.smashing.app.domain.model.Notification
+import com.smashing.app.data.model.notification.Notification
 
 interface NotificationRepository {
