+++ app/src/main/java/com/smashing/app/presentation/notice/NoticeContract.kt
@@ -2,7 +2,7 @@ package com.smashing.app.presentation.notice
 
 import androidx.compose.runtime.Immutable
 import com.smashing.app.data.model.cursor.Cursor
-import com.smashing.app.domain.model.Notification
+import com.smashing.app.data.model.notification.Notification
 import com.smashing.app.presentation.matching.type.MatchingType
 import com.smashing.app.presentation.notice.model.NoticeChangeSportUiModel
 import kotlinx.collections.immutable.ImmutableList
